requirements.txt
streamlit==1.28.0
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
folium==0.14.0
streamlit-folium==0.15.0
opencv-python==4.8.1.78
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
earthengine-api==0.1.384
geemap==0.30.0

rasterio==1.3.9
geopandas==0.14.1.80
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
scipy==1.11.4
scikit-learn==1.4.0
sentinelhub==3.9.1
python-dotenv==1.0.0
geopy>=2.3.0

reportlab>=4.0.0